fix(#7199): split top-comment tests out of eotest to satisfy qulice file length check
